Move ScaleLatentsInvocation and ResizeLatentsInvocation to their own file. No functional changes.
